refactor: decompose the compiler into per-subclass JIT twins in core/jit.ts
